[RDF] Move Python-related docs into PyROOT box
